Move & clean up model fetchers
