+++ src/api/providers/requesty.ts
@@ -1,6 +1,9 @@
-import { OpenAiHandler, OpenAiHandlerOptions } from "./openai"
+import axios from "axios"
+
 import { ModelInfo, requestyModelInfoSaneDefaults, requestyDefaultModelId } from "../../shared/api"
-import { ApiStream, ApiStreamUsageChunk } from "../transform/stream"
+import { parseApiPrice } from "../../utils/cost"
+import { ApiStreamUsageChunk } from "../transform/stream"
+import { OpenAiHandler, OpenAiHandlerOptions } from "./openai"
 
 export class RequestyHandler extends OpenAiHandler {
 	constructor(options: OpenAiHandlerOptions) {
@@ -37,4 +40,41 @@ export class RequestyHandler extends OpenAiHandler {
 			cacheReadTokens: usage?.cache_read_input_tokens,
 		}
 	}
+}
+
+export async function getRequestyModels({ apiKey }: { apiKey?: string }) {
+	const models: Record<string, ModelInfo> = {}
+
+	if (!apiKey) {
+		return models
+	}
+
+	try {
+		const config: Record<string, any> = {}
+		config["headers"] = { Authorization: `Bearer ${apiKey}` }
+
+		const response = await axios.get("https://router.requesty.ai/v1/models", config)
+		const rawModels = response.data.data
+
+		for (const rawModel of rawModels) {
+			const modelInfo: ModelInfo = {
+				maxTokens: rawModel.max_output_tokens,
+				contextWindow: rawModel.context_window,
+				supportsImages: rawModel.support_image,
+				supportsComputerUse: rawModel.support_computer_use,
+				supportsPromptCache: rawModel.supports_caching,
+				inputPrice: parseApiPrice(rawModel.input_price),
+				outputPrice: parseApiPrice(rawModel.output_price),
+				description: rawModel.description,
+				cacheWritesPrice: parseApiPrice(rawModel.caching_price),
+				cacheReadsPrice: parseApiPrice(rawModel.cached_price),
+			}
+
+			models[rawModel.id] = modelInfo
+		}
+	} catch (error) {
+		console.error(`Error fetching Requesty models: ${JSON.stringify(error, Object.getOwnPropertyNames(error), 2)}`)
+	}
+
+	return models
 }
